iqe-host-inventory-plugin/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatchling",
]

[project]
name = "iqe-host-inventory-plugin"
version = "25.99.1.dev1"
readme = "README.md"
maintainers = [
  { name = "Adarsh Dubey", email = "addubey@redhat.com" },
  { name = "Frantisek Stavela", email = "fstavela@redhat.com" },
  { name = "Mike Sager", email = "msager@redhat.com" },
  { name = "Zhamilya Abikenova", email = "zabikeno@redhat.com" },
]
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "click",
  "faker",
  "iqe-core>=24.11.29.1",
  "iqe-export-service-plugin",
  "iqe-ingress-plugin",
  "iqe-kessel-sync-plugin",
  "iqe-mq-plugin>=24.11.29",
  "iqe-notifications-plugin",
  "iqe-rbac-plugin",
  "ocdeployer",
  "psycopg2-binary",
  "pytest",
  "pytest-html",
  "pytest-lazy-fixtures",
  "requests",
  "sqlalchemy",
]
entry-points."iqe.application_plugins".host_inventory = "iqe_host_inventory:ApplicationHostInventory"
entry-points."iqe.click_commands".host_inventory = "iqe_host_inventory.grafted_commands:host_inventory_group"
entry-points.pytest11.iqe_host_inventory_plugin = "iqe_host_inventory.pytest_plugin"

[tool.hatch]
build.packages = [ "iqe_host_inventory", "iqe_host_inventory_api", "iqe_host_inventory_api_v7" ]
build.targets.sdist.include = [ "/iqe_host_inventory", "/iqe_host_inventory_api", "/iqe_host_inventory_api_v7" ]
build.targets.wheel.packages = [ "iqe_host_inventory", "iqe_host_inventory_api", "iqe_host_inventory_api_v7" ]
envs.default.python = "3.9"
envs.default.env-vars.PIP_INDEX = "https://nexus.corp.redhat.com/repository/cqt-pypi/pypi"
envs.default.env-vars.PIP_INDEX_URL = "https://nexus.corp.redhat.com/repository/cqt-pypi/simple"
# enable comms with nexus and stage CRC
envs.default.env-vars.REQUESTS_CA_BUNDLE = "/etc/pki/tls/certs/ca-bundle.crt"

[tool.ruff]
line-length = 99
preview = true
lint.extend-select = [
  "B",
  "C4",
  "C90",
  "E",
  "F",
  "FURB",
  "I",
  "RUF",
  "SIM",
  "UP",
  "W",
]
lint.extend-ignore = [
  "RUF045", # dataclass assignment without annotation
  "RUF067", # non-empty-init-module - intentional for plugin architecture
  "SIM117", # merge stacked with statements
  "SIM118", # replace dict.keys iteration with dictkeys set operations
]
lint.per-file-ignores."iqe_host_inventory_api*/*.py" = [
  "B904",
  "BLE001",
  "C901",
  "E501",
  "E721",
  "FURB103",
  "FURB171",
  "PERF102",
  "PERF401",
  "PLR1704",
  "RUF001",
  "RUF002",
  "RUF005",
  "RUF012",
  "RUF027",
  "RUF036",
  "RUF039",
  "RUF052",
  "S110",
  "SIM101",
  "SIM105",
  "SIM108",
  "TRY002",
  "TRY004",
  "TRY201",
  "TRY203",
  "UP042",
  "UP046",
  "YTT202",
]
lint.extend-safe-fixes = [
  "C402",
  "C405",
  "C408",
  "C409",
  "C417",
  "C419",
  "SIM103",
  # note - on regen with the old openap apis,
  # its sometimes necesary to delete some noqa comments for the rules to work propperly
  "UP006",
  "UP007",
  "UP008",
  "UP030",
  "UP031",
  "UP032",
]
lint.isort.force-single-line = true
